+++ lib/marv/generator.rb
@@ -22,7 +22,6 @@ module Marv
         ['assets', 'images'],
         ['assets', 'javascripts'],
         ['assets', 'stylesheets'],
-        ['assets', 'lib'],
 
         ['functions'],